+++ lithops/serverless/backends/gcp_functions/config.py
@@ -35,7 +35,7 @@ AVAILABLE_PY_RUNTIMES = {'3.7': 'python37', '3.8': 'python38', '3.9': 'python39'
 
 USER_RUNTIMES_PREFIX = 'lithops.user_runtimes'
 
-REQ_PARAMS = ('project_name', 'service_account', 'credentials_path', 'region')
+REQ_PARAMS = ('project_name', 'service_account', 'region')
 
 DEFAULT_CONFIG_KEYS = {
     'runtime_timeout': 300,  # Default: 5 minutes
@@ -55,6 +55,7 @@ google-cloud
 google-cloud-storage
 google-cloud-pubsub
 google-auth
+google-api-python-client
 certifi
 chardet
 docutils
@@ -85,10 +86,12 @@ def load_config(config_data=None):
             msg = f"{param} is mandatory under 'gcp' section of the configuration"
             raise Exception(msg)
 
-    config_data['gcp']['credentials_path'] = os.path.expanduser(config_data['gcp']['credentials_path'])
+    if 'credentials_path' not in config_data['gcp']:
+        if 'GOOGLE_APPLICATION_CREDENTIALS' in os.environ:
+            config_data['gcp']['credentials_path'] = os.environ.get('GOOGLE_APPLICATION_CREDENTIALS')
 
-    if not os.path.isfile(config_data['gcp']['credentials_path']):
-        raise Exception(f"Credentials file {config_data['gcp']['credentials_path']} not found")
+    if 'credentials_path' in config_data['gcp']:
+        config_data['gcp']['credentials_path'] = os.path.expanduser(config_data['gcp']['credentials_path'])
 
     for key in DEFAULT_CONFIG_KEYS:
         if key not in config_data['gcp_functions']: