+++ lithops/storage/backends/gcp_storage/gcp_storage.py
@@ -34,10 +34,12 @@ TIMEOUT = 5
 class GCPStorageBackend:
     def __init__(self, gcp_storage_config):
         logger.debug("Creating GCP Storage client")
-        self.credentials_path = gcp_storage_config['credentials_path']
+        self.credentials_path = gcp_storage_config.get('credentials_path')
         try:  # Get credenitals from JSON file
             self.client = storage.Client.from_service_account_json(self.credentials_path)
+            logger.debug(f'Getting GCP credentials from {self.credentials_path}')
         except Exception:  # Get credentials from gcp function environment
+            logger.debug(f'Getting GCP credentials from the environment')
             self.client = storage.Client()
         msg = STORAGE_CLI_MSG.format('Google Cloud Storage')
         logger.info("{}".format(msg))